Update wed night

+++ src/model/elements/Fixture.java
@@ -4,12 +4,12 @@ package model.elements;
  * The Fixture class represents immovable objects in the game.
  */
 public class Fixture extends GameElements {
-  private int weight;
+  private double weight;
 
   /**
    * Constructor: Initializes the name, description, and weight of the fixture.
    */
-  public Fixture(String name, String description, int weight) {
+  public Fixture(String name, String description, double weight) {
     super(name, description);  // Call the constructor of the superclass
     this.weight = weight;
   }
@@ -17,7 +17,7 @@ public class Fixture extends GameElements {
   /**
    * Gets the weight of the fixture.
    */
-  public int getWeight() {
+  public double getWeight() {
     return weight;
   }
